requirements.txt
-e .
pytest

aws_cdk.aws_apigateway
